test: add unit tests for removeEvent

+++ src/main/java/pl/envelo/moovelo/controller/event/EventController.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,6 +17,7 @@ import pl.envelo.moovelo.controller.dto.event.DisplayEventResponseDto;
 import pl.envelo.moovelo.controller.dto.event.EventListResponseDto;
 import pl.envelo.moovelo.controller.mapper.EventListResponseMapper;
 import pl.envelo.moovelo.controller.mapper.event.EventMapper;
+import pl.envelo.moovelo.entity.actors.BasicUser;
 import pl.envelo.moovelo.entity.events.CyclicEvent;
 import pl.envelo.moovelo.entity.events.Event;
 import pl.envelo.moovelo.entity.events.ExternalEvent;
@@ -55,8 +58,18 @@ public class EventController {
 
     @DeleteMapping("/events/{eventId}")
     @PreAuthorize("hasRole('BASIC_USER')")
-    public ResponseEntity removeEventById(@PathVariable long eventId) {
+    public ResponseEntity removeEventById(@PathVariable long eventId) throws IllegalAccessException {
         log.info("EventController - removeEventById() - eventId = {}", eventId);
+
+        //TODO: Do sprawdzenia przy spring security
+        /*Event event = eventService.getEventById(eventId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        BasicUser basicUser = (BasicUser) authentication.getPrincipal();
+
+        if (!event.getEventOwner().getUserId().equals(basicUser.getId())) {
+            throw new IllegalAccessException("You do not have permission to modify the chosen event");
+        }*/
+
         eventService.removeEventById(eventId);
         log.info("EventController - removeEventById() - event with eventId = {} removed", eventId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
